Check if working path is normal path before delete

+++ php/elFinderVolumeDriver.class.php
@@ -1755,7 +1755,9 @@ abstract class elFinderVolumeDriver {
 		
 		$work_path = $this->getWorkFile($path);
 		if (!is_writable($work_path)) {
-			is_file($work_path) && @unlink($work_path);
+			if ($path !== $work_path) {
+				is_file($work_path) && @unlink($work_path);
+			}
 			return false;
 		}
